d2 generator will actually pick up overrides
This was caused by two issues, one being that the generator only checked
for certain kinds of key types in the schema as candidates to make
references for.

The other issue was that Go doesn't consider a mutation of a ranging
value across a slice to mutate the original slice.
This could likey have been done using pointers instead, but the process
works fine with the added copy-ing steps I've added