Simplify mysql Implementation
This sinks the DB connection details further down into the hex.
By disregarding the connection details the main function
now no longer needs to be concerned with DB details.

It also simplifies the exposed API for the mysql module.

This will hopefully make it simpler for future extensions with other dbs